Refactor state management to use Zustand and remove useStory hook
