+++ app/(tabs)/storyscreen/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { View, Text } from 'react-native';
 import Layout from '../../../components/Layout/Layout';
-import { useStory } from '@/hooks/useStory';
+import { useConversationStore } from '@/src/stores/conversationStore';
 import ResponseList from '@/components/ResponseList/ResponseList';
 import GenerateButton from '@/components/GenerateButton/GenerateButton';
 import StoryContent from '@/components/StoryContent/StoryContent';
@@ -10,18 +10,14 @@ import { s } from '../../../pages/StoryScreen/StoryScreen.style';
 
 const StoryScreen = () => {
   const {
-    storyState,
-    story,
-    handleConversationComplete,
-    generateStoryWithImages,
-  } = useStory();
-
-  const {
-    question,
+    currentQuestion: question,
     responses,
     isGenerating,
     conversationComplete,
-  } = storyState;
+    story,
+    handleConversationComplete,
+    generateStoryWithImages,
+  } = useConversationStore();
 
   return (
     <Layout>
