+++ metro.config.js
@@ -9,4 +9,8 @@ config.resolver.extraNodeModules = {
   'react-dom': require.resolve('react-dom'),
 };
 
+// Add resolver configuration for ES modules compatibility
+config.resolver.unstable_enablePackageExports = false;
+config.resolver.unstable_conditionNames = ['react-native', 'browser', 'require'];
+
 module.exports = config;
